move to jsk_interactive_marker and modify spacenav rotate
add GetPose.srv
