+++ jsk_interactive_markers/jsk_interactive_marker/launch/transformable_server_sample.launch
@@ -3,10 +3,10 @@
 
   <param name="torus_udiv" value="100"/>
   <param name="torus_vdiv" value="100"/>
-  <node pkg="jsk_transformable_interactive_marker" type="transformable_server_sample" name="server_sample">
+  <node pkg="jsk_interactive_marker" type="transformable_server_sample" name="server_sample">
   </node>
 
-  <node pkg="rviz" type="rviz" name="rviz" args=" -d $(find jsk_transformable_interactive_marker)/config/transformable_server_sample.rviz"/>
+  <node pkg="rviz" type="rviz" name="rviz" args=" -d $(find jsk_interactive_marker)/config/transformable_server_sample.rviz"/>
 
 
   <!-- For KORG NANO KONTROL -->
@@ -14,7 +14,7 @@
     <remap from="joy" to="korg_joy"/>
   </node>
 
-  <node pkg="jsk_transformable_interactive_marker" type="transformable_joy_configure.py" name="transformable_joy_configure">
+  <node pkg="jsk_interactive_marker" type="transformable_joy_configure.py" name="transformable_joy_configure">
     <remap from="input_joy" to="korg_joy"/>
 
     <!-- the values are calclated with below equation -->
@@ -37,7 +37,7 @@
   <node pkg="spacenav_node" type="spacenav_node" name="spacenav">
   </node>
 
-  <node pkg="jsk_transformable_interactive_marker" type="transformable_spacenav_configure.py" name="transformable_spacenav_configure">
+  <node pkg="jsk_interactive_marker" type="transformable_spacenav_configure.py" name="transformable_spacenav_configure">
     <remap from="input_joy" to="/spacenav/joy"/>
 
     <!-- choose using both at same time or using separately -->
