+++ jsk_interactive_markers/jsk_interactive_marker/src/transformable_object.cpp
@@ -1,6 +1,6 @@
-#include <jsk_transformable_interactive_marker/transformable_object.h>
+#include <jsk_interactive_marker/transformable_object.h>
 
-using namespace jsk_transformable_interactive_marker;
+using namespace jsk_interactive_marker;
 
 TransformableObject::TransformableObject(){
   ROS_INFO("Init TransformableObject");
@@ -88,24 +88,47 @@ void TransformableObject::addPose(geometry_msgs::Pose msg){
   pose_.position.z += msg.position.z;
   float tmp_x = pose_.orientation.x, tmp_y = pose_.orientation.y,
     tmp_z = pose_.orientation.z, tmp_w = pose_.orientation.w;
+
+
+  // pose_.orientation.w =
+  //   tmp_w * msg.orientation.w -
+  //   tmp_x * msg.orientation.x -
+  //   tmp_y * msg.orientation.y -
+  //   tmp_z * msg.orientation.z;
+  // pose_.orientation.x =
+  //   tmp_y * msg.orientation.z -
+  //   tmp_z * msg.orientation.y +
+  //   tmp_y * msg.orientation.x +
+  //   tmp_x * msg.orientation.w;
+  // pose_.orientation.y =
+  //   tmp_z * msg.orientation.x -
+  //   tmp_x * msg.orientation.z +
+  //   tmp_w * msg.orientation.y +
+  //   tmp_y * msg.orientation.w ;
+  // pose_.orientation.z =
+  //   tmp_x * msg.orientation.y -
+  //   tmp_y * msg.orientation.x +
+  //   tmp_w * msg.orientation.z +
+  //   tmp_z * msg.orientation.w;
+
   pose_.orientation.w =
-    tmp_w * msg.orientation.w -
-    tmp_x * msg.orientation.x -
-    tmp_y * msg.orientation.y -
-    tmp_z * msg.orientation.z;
+    msg.orientation.w * tmp_w -
+    msg.orientation.x * tmp_x -
+    msg.orientation.y * tmp_y -
+    msg.orientation.z * tmp_z;
   pose_.orientation.x =
-    tmp_y * msg.orientation.z -
-    tmp_z * msg.orientation.y +
-    tmp_y * msg.orientation.x +
-    tmp_x * msg.orientation.w;
+    msg.orientation.y * tmp_z -
+    msg.orientation.z * tmp_y +
+    msg.orientation.y * tmp_x +
+    msg.orientation.x * tmp_w;
   pose_.orientation.y =
-    tmp_z * msg.orientation.x -
-    tmp_x * msg.orientation.z +
-    tmp_w * msg.orientation.y +
-    tmp_y * msg.orientation.w ;
+    msg.orientation.z * tmp_x -
+    msg.orientation.x * tmp_z +
+    msg.orientation.w * tmp_y +
+    msg.orientation.y * tmp_w ;
   pose_.orientation.z =
-    tmp_x * msg.orientation.y -
-    tmp_y * msg.orientation.x +
-    tmp_w * msg.orientation.z +
-    tmp_z * msg.orientation.w;
+    msg.orientation.x * tmp_y -
+    msg.orientation.y * tmp_x +
+    msg.orientation.w * tmp_z +
+    msg.orientation.z * tmp_w;
 }
