+++ jsk_interactive_markers/jsk_interactive_marker/src/transformable_server_sample.cpp
@@ -1,7 +1,7 @@
 #include <ros/ros.h>
-#include <jsk_transformable_interactive_marker/transformable_interactive_server.h>
+#include <jsk_interactive_marker/transformable_interactive_server.h>
 
-using namespace jsk_transformable_interactive_marker;
+using namespace jsk_interactive_marker;
 
 int main(int argc, char** argv)
 {
