+++ jsk_interactive_markers/jsk_interactive_marker/srv/GetPose.srv
@@ -1,0 +1,3 @@
+string target_name
+---
+geometry_msgs/Pose pose
