--- jsk_interactive_markers/jsk_transformable_interactive_marker/package.xml
@@ -1,24 +1,0 @@
-<?xml version="1.0"?>
-<package>
-  <name>jsk_transformable_interactive_marker</name>
-  <version>0.0.0</version>
-  <description>The jsk_transformable_interactive_marker package</description>
-
-  <maintainer email="inagaki@todo.todo">inagaki</maintainer>
-
-  <license>MIT</license>
-
-  <buildtool_depend>catkin</buildtool_depend>
-  <build_depend>interactive_markers</build_depend>
-  <build_depend>roscpp</build_depend>
-  <build_depend>std_msgs</build_depend>
-  <build_depend>geometry_msgs</build_depend>
-  <build_depend>tf</build_depend>
-  <build_depend>visualization_msgs</build_depend>
-  <run_depend>interactive_markers</run_depend>
-  <run_depend>roscpp</run_depend>
-  <run_depend>std_msgs</run_depend>
-  <run_depend>geometry_msgs</run_depend>
-  <run_depend>tf</run_depend>
-  <run_depend>visualization_msgs</run_depend>
-</package>
